Refactor FilterConfig into separate controllers; add more tests
